collision optimization & rename BikePart to Part
