+++ class/entity/Part.js
